+++ src/lib.rs
@@ -100,7 +100,9 @@ pub mod samples;
 pub mod util;
 
 pub use boxes::{BoxHeader, BoxKey, BoxRef, FourCC, NodeKind};
-pub use parser::{ParseIssue, parse_boxes, parse_boxes_tolerant, parse_children, read_box_header};
+pub use parser::{
+    MAX_BOX_DEPTH, ParseIssue, parse_boxes, parse_boxes_tolerant, parse_children, read_box_header,
+};
 pub use registry::{
     BoxValue, Co64Data, CttsData, CttsEntry, HdlrData, MdhdData, Registry, SampleEntry, StcoData,
     StructuredData, StscData, StscEntry, StsdData, StssData, StszData, SttsData, SttsEntry,